Bug Fix: Bug # People donated count is not increasing.
Problem: From old code for displaying people donated count, calculations are calculated as:for example lets assume total people donated is 910.then 910/50 = 18.2so we are increasing count in the multiples of 18 until last multiple is less than 910.In this case 18 * 50 = 900. So it is displaying 900 on home page. by this logic people donated count increases on home page when total donation amount crosses 918, 936, 954, etc.

Solution:
- Now we have modified counter logic and displayed our exact people donated count.